+++ app/services/timeline/month_summary.rb
@@ -150,19 +150,29 @@ module Timeline
     end
 
     def track_seconds
-      @track_seconds ||= @user.scoped_tracks
-                              .where(start_at: month_range)
-                              .group(date_sql_expr('tracks.start_at'))
-                              .sum(:duration)
-                              .transform_keys(&:to_s)
+      @track_seconds ||= track_day_attributions[:seconds]
     end
 
     def track_count
-      @track_count ||= @user.scoped_tracks
-                            .where(start_at: month_range)
-                            .group(date_sql_expr('tracks.start_at'))
-                            .count
-                            .transform_keys(&:to_s)
+      @track_count ||= track_day_attributions[:count]
+    end
+
+    def track_day_attributions
+      @track_day_attributions ||= begin
+        seconds = Hash.new(0.0)
+        count = Hash.new(0)
+        overlapping_tracks.find_each do |track|
+          TrackDayShares.for(track, tz).each do |day, fraction|
+            seconds[day.to_s] += track.duration.to_f * fraction
+            count[day.to_s] += 1
+          end
+        end
+        { seconds: seconds, count: count }
+      end
+    end
+
+    def overlapping_tracks
+      @user.scoped_tracks.where('start_at <= ? AND end_at >= ?', month_range.last, month_range.first)
     end
 
     def visit_status_label(status)